Fix uncaught exception in os.symlink
There may be a case when user have a module build directory with compile_commands.json file, and at the same time user deleted the module source directory. In this case KDECmake::configureInternal even in pretend mode will attempt to create a symlink to that file in the (non-existing) source directory.