vscode: don't use CMakePresets.json files
If there is a CMakePresets.json file in the project, the CMakeTools
extension will use it by default. This is an issue because this
configuration will not use the environment variables set by the
other config files like `settings.json`.

This may be resolved in the future by this open issue:
https://github.com/microsoft/vscode-cmake-tools/issues/3578

Also possibly if this feature were added:
https://invent.kde.org/sdk/kde-builder/-/issues/52

By adding this setting the extension will use the CMakeLists.txt along
with the configurations we generate, which is what we want so projects
are configured automatically.