Close monitor-to-build ipc in proper place
Do close monitorToBuildIPC only when we have done with any logger messages. Previously it was closed before we may want to send error message. Most likely because of this issue, the Debug.error() was handled differently (printed to stderr instead of passing to print_clr as all other methods do) in kdesrc-build.