Print error log path immediately after fail when stop-on-failure is false
When you build long list of modules, or some of the modules have long build time, and you set stop-on-failure to false, and when some of the modules failed, you have to wait until all modules are built, to be able to see paths to error log files for modules.

This commit prints the error log of failed module right after fail, so developer can already read the log, while other modules are compiling.

Fixes #39