Ignore command line selectors when using --resume or --rebuild-failures
kdesrc-build assumed that user do not pass any selectors when using --resume or --rebuild-failures options. In fact, it prepended the selectors from command line (assumed empty) with the selectors from global persistent options "resume-list" or "last-failed-module-list".

While in kde-builder this was needed to be ported as `selectors[:0] = moduleList.split(", ")`, I decided to handle it better.

The user ofter would want to append --resume option for the previous command, and they may forget to remove selector, effectively shooting in the leg (getting unexpected list of modules).
And it is better to _allow_ such appending for convenience, but just ignore the selectors presented in command line.
If using --debug option, user may see the message that they were ignored.