Remove BuildSystemQt6 class
The only difference in BuildSystemQt6 in comparison with BuildSystemKDECmake is that BuildSystemQt6 set qt-install-dir to install-dir before calling prepare_module_build_environment().

This, however, is not needed. In the prepare_module_build_environment(), before setting variables using path from qt-install-dir, there is a check `qt_installdir != prefix` (where prefix is read from `install-dir`), which will never be true (i.e. the qt-install-dir was set to be equal to install-dir just in the overriden method and just before calling super class method).