fix: generate IDE configs _before_ running cmake, to allow debugging failing cmake
In case cmake configure fails, the module is marked as failed to build,
and the rest of commands are not executed by kde-builder. And as the
user has not yet configs, and wants to run cmake script in their IDE,
they need to create ide configuration manually, which makes the
kde-builder generation function useless in this case.

This commit fixes this.