fix: Allow use of third-party dependencies in dependencies-data
The previous commit introduced a bug. It did exit application as long as non-kde project was referenced.
But it was intended to be ignored for third-party dependencies.

This commit still keeps the idea of previous commit: it does not show ugly traceback, and instead prints
the simple error message. We use a new exception type for this (UnknownKdeProjectException), and will
catch it and treat as appropriate in the kde-builder script.