refactor: Remove unneeded install-after-build option
The --(no-)install-after-build option was confusing because it duplicated --(no-)install option.
In build() function, the install() function was invoked unless install-after-build is set to false. But
the user may use no-install option, which the build() function did not respected. This commit fixes that.

Also, install-after-build was documented incorrectly (it was Bool type, not String).
Also, no-install option already filters out the install phase.