fix: Support using pretending mode at first run (when no git config, and no metadata)
When user has just installed kde-builder, the metadata is not yet downloaded.
For metadata module, we use a repository address with "kde:" alias in it,
because we want to support the git (not https) push protocol
(for the cases when user wants to contribute to metadata in the future).
verify_git_config() function does setting the alias in git config.
And as we ignore the pretending mode to download metadata for the first time,
we also must ignore pretending mode for the alias setting.