refactor: Remove the sticky options concept
The idea was that you can set a temporary "overriding"
option for normal one, like "#source-dir", which would
be returned in preference of "source-dir".

Additionally, if build_context (a global options object)
had some option starting with "#", that would always
be returned in preference of project's "normal" option.

This was not really used in the code.