Remove needs_builddir_hack() function from qmake5/qmake6 build systems
The "build dir hack" means that build directory is symlinked to
 source directory, so they actually become the same directory.

Initially, (in kdesrc-build equivalent functions and classes) the Qt5 modules
were built with BuildSystemQmake. Then the BuildSystemQt5 was developed, which
stopped using the build dir hack: it is inherited from BuildSystem (generic build
system), which always returns False.

In the needs_builddir_hack() implementation, the True value was only returned
when the "repository" option was containing "qt.io".
The only two projects that kde-builder builds that use qmake currently are:
- libaccounts-qt (repository: "https://gitlab.com/accounts-sso/libaccounts-qt.git")
- signond (repository: "https://gitlab.com/accounts-sso/signond.git")
Thus the build dir hack is not applied to them either.

So, this commit removes this useless function.