fix: Display "possible solution" text when failed to configure project
The text "Possible solution: Install the build dependencies for the
modules" was never shown.

The re.match searches in the beginning of the string, while re.search
searches in the whole string. We need re.search.

Also, show only failed projects that failed at configure step, but not
every failed project.