Always build when holding work branches
Currently, on hold-work-branches=true and build-when-unchanged=false,
build is skipped on a work branch (unless --build-when-unchanged is
specified in the command line).

However, for me the most frequent usage of kde-builder is exactly
the above case, i.e. building a work branch. It would be nice to
not having to specify --build-when-unchanged everytime on a work
branch.