fix: Make repo-metadata path on disk independent of directory-layout value
The default "directory-layout" option value is "flat".
The expected location for repo-metadata repo on disk is "~/.local/state/sysadmin-repo-metadata".

Now if user changes their "directory-layout" option in their config or in command line,
to the value "invent" or "metadata", then projects' source directories are laid under subfolders.

But repo-metadata source directory path should be independent of this option.