remove: Do not use dimmed color in printed messages
It was only used in messages in pretending mode (and in one place when printing
log file path). It was presented as grey color (with Breeze color theme in Konsole
profile, which is used by default), or forced to blue in tty.
We can instead print them in normal color.