feat: Support "kde-projects" magic value in project nodes
Previously, the magic value "kde-projects" was only allowed to be set
in the group nodes (and override nodes), but not in project nodes.
So it was impossible to declare options for KDE project alone (not
inside a group). It required extra effort to either specify repository
path or wrapping project inside some group just to get that magic
repository resolution and thus avoiding specifying it manually.

In other words, such declaration was impossible:

```yaml
  project elisa:  # <- real kde project name
    # <- not specifying "repository" here
    cmake-options: -DBUILD_WITH_QT6=ON
```

And the workaround was to either specify the repository option like this:

```yaml
  project elisa:
    repository: kde:multimedia/elisa.git  # <- could not omit this option
    cmake-options: -DBUILD_WITH_QT6=ON
```

or wrap in unneeded group like this:

```yaml
  group elisa:
    repository: kde-projects  # <- magic was supported in group
    cmake-options: -DBUILD_WITH_QT6=ON
    use-projects:
      - elisa
```

And specifying options for KDE projects like that was inconvenient.

This commit allows to use "kde-projects" magic value in "repository"
option of project nodes.

Also, even this can now be omitted for known kde projects, because
this value is now a default (both for projects and groups).

Implementation detail: the value of "repository" that is set in the
group is now inherited by projects as all other normal options.
Previously, it was resolved when expanding groups. But now, it will
be resolved for the projects in another internal option
"#resolved-repository".

Fixes #138