refactor: Remove "#selected-by" and "#found-by" internal options
They were used only in _module_is_needed(), which itself only influenced
the message text when the requested branch does not exist in repo.

"#found-by" was absolutely useless and is not part of the metadata.
"#selected-by" was not useful, as in the place where that can potentially
be used - explicit_kdeproject_selectors - we do not use it.

Also, remove _module_is_needed() itself.