fix: Prevent ignoring branch-group for KDE projects in specific circumstances
If some kde project appeared in build-configs as bare "project" node, and that happened
before any group/override mentioned it, when it was created, its scm_type object was not
initialized. That lead to that when it was attempted to be source-updated, its scm_type
was initialized as "git" (instead of correct "proj"), which itself lead to that it was
appeared to the user as third-party. And the branch-group from repo-metadata was ignored
for it.