fix: Avoid creating project build dir when --src-only is used
The update process created a file `~/kde/build/<proj_name>/kde-builder.env`.
This file did not contain any variables because we do not set them
in the update phase. And it was normally rewritten by build phase.

But in case user nuked their build dir, and wanted to just update
all project sources, they do not expect that the build directories
will be created.