fix: Remove faulty optimization for skipping installation
In kdesrc-build, there was an attempt to add optimization for skipping
installation in case the ninja build tool "had no work to do" during
build. That feature was requested in https://invent.kde.org/sdk/kdesrc-build/-/issues/92.

This optimization was introduced by commit:
https://invent.kde.org/sdk/kdesrc-build/-/commit/8c693eeef39466bfd3dc3b7132b3200884236aa5

A new problem was that even with "--refresh-build", it was not possible
to bypass the unwanted "optimization" triggering.
CCBUG: 456593

Later, the check for presence of "--refresh-build" was added:
https://invent.kde.org/sdk/kdesrc-build/-/commit/7e1a75e66d5b6ac4fc0db3e828d13728cd04f83c

However, a problem still remained. Sometimes, users want to update some
project files, which do not affect "ninja work", and reinstall the
project. And user may not want to use the "--refresh-build", because
it is very expensive. See this issue:
https://invent.kde.org/sdk/kdesrc-build/-/issues/107

In kdesrc-build, the optimization was half-incorrectly disabled in
https://invent.kde.org/sdk/kdesrc-build/-/commit/fb7c8f06a082ec676798434cc33e8f7523f6bdab.
It was done by not allowing "work_done" variable be set to false
(except the bug when using --debug, which was fixed in kde-builder).

Also, there was a problem with making a decision if installation is needed,
depending on the presence of "last-install-rev" in persistent options.
This was not entirely correct. The user may have some local changes,
not yet committed (i.e. having the same install-revision number).
And this check forced user to skip reinstallation of the project.

So, there is no point in such "optimization". At least, with such implementation.
So remove it now.