"Asynchronize" the join() by running it in a thread
This should decrease idle time.
The sleep(1) always waited a second even if the process finished
earlier. Decreasing the sleep time would also decrease idle time,
but it's cleaner (if not necessarily the most efficient possible
way) to just "asynchronize" the join() by running it in a worker
thread (pool).