"Asynchronize" the Queue.get() by running it in a thread
Eliminating the waiting means a large speedup for the processing of
projects with little to do, proportionally much less otherwise.
This also makes the displayed build process percentage update much
more rapidly, which is a matter of taste I guess.

There was a problem with kde-builder not terminating after pressing
Ctrl-C because lines_queue.get() got stuck waiting forever after the
sending side of the queue was killed, preventing the worker thread
it was executing in from quitting cleanly (Python doesn't seem to
want to kill busy worker threads).

The fix is to pass a timeout to get(), which wastes a bit of CPU
time like the previous asyncio.sleep() based approach, but it
*doesn't* wait for longer than necessary if there *are* messages in
the queue.
There is probably a better way. If the better way involves signal
handlers, I'm not sure if it's actually better though.