Introduce a "main function" as per the usual Python convention
Without it, Python never considers the runtime as fully
initialized and throws an exception when trying to use the
fork_server start method of the multiprocessing module.
fork_server is the default in Python 3.14 and strongly recommended
even before because of correctness issues that the fork method has
with multithreading.