doc: Correctly document the nice range
The default value of niceness in Linux is 0.
It can be increased up to 19.
Negative values are not applicable for kde-builder,
because it is not intended to be run as root.
There is a special note in the doc here, that it is
not a good idea anyway, even if running as root.