fix: Filter metadata-ignored projects in the generate_module_list()
The "query" run mode was using half-unfiltered list of Module objects,
which caused that the projects ignored in repo-metadata were still
appearing in query output list. But in normal ("build") run mode, they
were filtered correctly.
The final filtering was done not in the generate_module_list(), but
in the run_all_module_phases().
Now we will do it in one place.