fix: Disable abbrev in command line options
The "abbrev" function of argparse allows to use not fully typed
option name in case there is no ambiguity. For example, if we
have option "--list-installed", and no any other option that
starts with "--lis", then users can use "--list-inst", or
"--list-in", or "--list".
However, this makes unnecessary confusion about existence of
some options, such as "--list" in this example. Also, this
encourages a bad habit. Users should use short options, or
autocompletion, but not incorrect representation of options.